better naming for octopus layout
