+++ octopus_code/core/octopus/device.py
@@ -1,0 +1,6 @@
+
+# TODO: is there anything outside of the constructor that will actually be shared?
+class Device:
+    def __init__(self, control_queue, audio_stream_queue):
+        self.control_queue = control_queue
+        self.audio_stream_queue = audio_stream_queue
