+++ octopus_code/core/octopus/layouts/pixel.py
@@ -1,0 +1,4 @@
+class Pixel:
+    def __init__(self, location, color=(0,0,0)):
+        self.location = location
+        self.color = color
