+++ octopus_code/core/tests/testSuite.py
@@ -4,6 +4,8 @@ from core.tests.gentlemanOctopusTests import TestGentlemanOctopus
 
 import unittest
 
+#TODO Clean build
+
 if __name__ == '__main__':
     unittest.TestLoader().loadTestsFromTestCase(TestIntegrationTest)
     unittest.TestLoader().loadTestsFromTestCase(TestGentlemanOctopus)
